fix: signposting tasklist is shown in preview mode
