+++ app/client/src/pages/Editor/Explorer/Pages/index.tsx
@@ -40,6 +40,7 @@ import useResize, {
 } from "utils/hooks/useResize";
 import AnalyticsUtil from "utils/AnalyticsUtil";
 import { useLocation } from "react-router";
+import { toggleInOnboardingWidgetSelection } from "actions/onboardingActions";
 
 const ENTITY_HEIGHT = 36;
 const MIN_PAGES_HEIGHT = 60;
@@ -116,6 +117,7 @@ function Pages() {
         type: "PAGES",
         toUrl: navigateToUrl,
       });
+      dispatch(toggleInOnboardingWidgetSelection(true));
       history.push(navigateToUrl);
     },
     [location.pathname],
