+++ app/client/src/selectors/onboardingSelectors.tsx
@@ -6,9 +6,15 @@ import { AppState } from "reducers";
 import { createSelector } from "reselect";
 import { getUserApplicationsWorkspaces } from "./applicationSelectors";
 import { getWidgets } from "sagas/selectors";
-import { getActionResponses, getActions } from "./entitiesSelector";
+import {
+  getActionResponses,
+  getActions,
+  getCanvasWidgets,
+} from "./entitiesSelector";
 import { getSelectedWidget } from "./ui";
 import { GuidedTourEntityNames } from "pages/Editor/GuidedTour/constants";
+import { previewModeSelector } from "./editorSelectors";
+import { commentModeSelector } from "./commentsSelectors";
 
 // Signposting selectors
 export const getEnableFirstTimeUserOnboarding = (state: AppState) => {
@@ -37,6 +43,32 @@ export const getIsFirstTimeUserOnboardingEnabled = createSelector(
 
 export const getInOnboardingWidgetSelection = (state: AppState) =>
   state.ui.onBoarding.inOnboardingWidgetSelection;
+
+export const getIsOnboardingWidgetSelection = (state: AppState) =>
+  state.ui.onBoarding.inOnboardingWidgetSelection;
+
+export const getIsOnboardingTasksView = createSelector(
+  getCanvasWidgets,
+  getIsFirstTimeUserOnboardingEnabled,
+  getIsOnboardingWidgetSelection,
+  previewModeSelector,
+  commentModeSelector,
+  (
+    widgets,
+    enableFirstTimeUserOnboarding,
+    isOnboardingWidgetSelection,
+    inPreviewMode,
+    inCommentMode,
+  ) => {
+    return (
+      Object.keys(widgets).length == 1 &&
+      enableFirstTimeUserOnboarding &&
+      !isOnboardingWidgetSelection &&
+      !inPreviewMode &&
+      !inCommentMode
+    );
+  },
+);
 
 // Guided Tour selectors
 export const isExploringSelector = (state: AppState) =>
